# Code/Server/docs/swagger.yaml
basePath: /v1
definitions:
...
  db.FixStatus:
    enum:
    - pending
    - planned
    - awaiting_owner_confirmation
    - awaiting_tenant_confirmation
    - fixed
    type: string
    x-enum-varnames:
    - FixStatusPending
    - FixStatusPlanned
    - FixStatusAwaitingOwnerConfirmation
    - FixStatusAwaitingTenantConfirmation
    - FixStatusFixed
  db.Priority:
    enum:
    - low
    - medium
    - high
    - urgent
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
    - PriorityUrgent
...
  models.DamageOwnerUpdateRequest:
    properties:
      fix_planned_at:
        type: string
      read:
        type: boolean
    type: object
  models.DamageRequest:
    properties:
      comment:
        type: string
      priority:
        $ref: '#/definitions/db.Priority'
      room_id:
        type: string
    required:
    - comment
    - priority
    - room_id
    type: object
  models.DamageResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      fix_planned_at:
        type: string
      fix_status:
        $ref: '#/definitions/db.FixStatus'
      fixed_at:
        type: string
      id:
        type: string
      lease_id:
        type: string
      picture_urls:
        items:
          type: string
        type: array
      priority:
        $ref: '#/definitions/db.Priority'
      read:
        type: boolean
      room_id:
        type: string
      room_name:
        type: string
      tenant_name:
        type: string
      updated_at:
        type: string
    type: object
  models.DamageTenantUpdateRequest:
    properties:
      comment:
        type: string
      priority:
        $ref: '#/definitions/db.Priority'
    type: object
...
  models.IdResponse:
    properties:
      id:
        type: string
    type: object
...
  utils.Error:
    properties:
      code:
        $ref: '#/definitions/utils.ErrorCode'
      error:
        type: string
    type: object
  utils.ErrorCode:
    enum:
    - invalid-password
    - cannot-fetch-user
    - user-not-found
    - cannot-create-user
    - no-claims
    - cannot-decode-user
    - missing-fields
    - missing-file
    - cannot-hash-password
    - email-already-exists
    - test-error
    - too-many-requests
    - invite-not-found
    - user-must-have-same-email-as-invite
    - invite-already-exists-for-email-or-property
    - lease-already-exists-for-tenant-and-property
    - property-not-found
    - property-is-not-yours
    - not-an-owner
    - not-a-tenant
    - user-already-exists-as-owner
    - property-already-exists
    - property-not-available
    - tenant-already-has-lease
    - no-active-lease
    - lease-not-found
    - cannot-end-non-current-lease
    - inventory-report-must-be-linked-to-current-lease
    - no-pending-lease
    - document-not-found
    - damage-not-found
    - damage-already-exists
    - cannot-update-fixed-damage
    - damage-already-fixed
    - failed-upload-file
    - failed-delete-file
    - bad-base64-string
    - property-picture-not-found
    - user-profile-picture-not-found
    - room-not-found
    - room-already-exists
    - room-not-in-this-property
    - furniture-not-found
    - furniture-already-exists
    - furniture-not-in-this-property
    - inventory-report-already-exists
    - cannot-modify-inventory-report
    - inventory-report-not-found
    - room-state-already-exists
    - furniture-state-already-exists
    - invalid-object-type
    - error-request-chatgpt-api
    - failed-send-email
    type: string
    x-enum-varnames:
    - InvalidPassword
    - CannotFetchUser
    - UserNotFound
    - CannotCreateUser
    - NoClaims
    - CannotDecodeUser
    - MissingFields
    - MissingFile
    - CannotHashPassword
    - EmailAlreadyExists
    - TestError
    - TooManyRequests
    - InviteNotFound
    - UserSameEmailAsInvite
    - InviteAlreadyExists
    - LeaseAlreadyExist
    - PropertyNotFound
    - PropertyNotYours
    - NotAnOwner
    - NotATenant
    - UserAlreadyExistsAsOwner
    - PropertyAlreadyExists
    - PropertyNotAvailable
    - TenantAlreadyHasLease
    - NoActiveLease
    - LeaseNotFound
    - CannotEndNonCurrentLease
    - InvReportMustBeCurrentLease
    - NoLeaseInvite
    - DocumentNotFound
    - DamageNotFound
    - DamageAlreadyExists
    - CannotUpdateFixedDamage
    - DamageAlreadyFixed
    - FailedUploadFile
    - FailedDeleteFile
    - BadBase64String
    - PropertyPictureNotFound
    - UserProfilePictureNotFound
    - RoomNotFound
    - RoomAlreadyExists
    - RoomNotInThisProperty
    - FurnitureNotFound
    - FurnitureAlreadyExists
    - FurnitureNotInThisProperty
    - InventoryReportAlreadyExists
    - CannotModifyInventoryReport
    - InventoryReportNotFound
    - RoomStateAlreadyExists
    - FurnitureStateAlreadyExists
    - InvalidObjectType
    - ErrorRequestChatGPTAPI
    - FailedSendEmail
host: localhost:3001
info:
  contact:
    email: martin.d-herouville@epitech.eu
    name: Mazettt
  description: This is the API used by the Immotep application.
  termsOfService: http://swagger.io/terms/
  title: Immotep API
  version: "1.0"
paths:
...
  /owner/properties/{property_id}/damages/:
    get:
      consumes:
      - application/json
      description: Get all damages of a property, optionally filtered by fixed status
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: 'Filter by fixed status (default: false)'
        in: query
        name: fixed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of damages
          schema:
            items:
              $ref: '#/definitions/models.DamageResponse'
            type: array
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: No active lease
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get property damages
      tags:
      - damage
...
  /owner/properties/{property_id}/leases/{lease_id}/damages/:
    get:
      consumes:
      - application/json
      description: Get all damages of a lease, optionally filtered by fixed status
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: 'Filter by fixed status (default: false)'
        in: query
        name: fixed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of damages
          schema:
            items:
              $ref: '#/definitions/models.DamageResponse'
            type: array
        "403":
          description: Lease not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: No active lease
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get lease damages
      tags:
      - damage
  /owner/properties/{property_id}/leases/{lease_id}/damages/{damage_id}/:
    get:
      consumes:
      - application/json
      description: Get a damage
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: Damage ID
        in: path
        name: damage_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Damage
          schema:
            $ref: '#/definitions/models.DamageResponse'
        "403":
          description: Lease not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Damage not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get damage
      tags:
      - damage
    put:
      consumes:
      - application/json
      description: Update damage on owner side. Owner can only update the read status
        and the fix planned date.
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: Damage ID
        in: path
        name: damage_id
        required: true
        type: string
      - description: Damage update request
        in: body
        name: damages
        required: true
        schema:
          $ref: '#/definitions/models.DamageOwnerUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated damage ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Damage not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Update damage for owner
      tags:
      - damage
  /owner/properties/{property_id}/leases/{lease_id}/damages/{damage_id}/fix/:
    put:
      consumes:
      - application/json
      description: Fix a damage for a tenant or owner. When both users have fixed
        the damage, it will be marked as fixed.
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: Damage ID
        in: path
        name: damage_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Fixed damage ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Lease not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Damage not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Fix damage for one user
      tags:
      - damage
...
  /tenant/leases/{lease_id}/damages/:
    get:
      consumes:
      - application/json
      description: Get all damages of a lease, optionally filtered by fixed status
      parameters:
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: 'Filter by fixed status (default: false)'
        in: query
        name: fixed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of damages
          schema:
            items:
              $ref: '#/definitions/models.DamageResponse'
            type: array
        "403":
          description: Lease not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: No active lease
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get lease damages
      tags:
      - damage
    post:
      consumes:
      - application/json
      description: Create a damage to a lease
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Lease ID or `current`
        in: path
        name: lease_id
        required: true
        type: string
      - description: Damages to create
        in: body
        name: damages
        required: true
        schema:
          $ref: '#/definitions/models.DamageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created damage ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields or bad base64 string
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: No active lease
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Create damage
      tags:
      - damage
  /tenant/leases/{lease_id}/damages/{damage_id}/:
    get:
      consumes:
      - application/json
      description: Get a damage
      parameters:
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: Damage ID
        in: path
        name: damage_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Damage
          schema:
            $ref: '#/definitions/models.DamageResponse'
        "403":
          description: Lease not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Damage not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get damage
      tags:
      - damage
    put:
      consumes:
      - application/json
      description: Update damage on tenant side. Tenant can only update comment, priority
        and add new pictures.
      parameters:
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: Damage ID
        in: path
        name: damage_id
        required: true
        type: string
      - description: Damage update request
        in: body
        name: damages
        required: true
        schema:
          $ref: '#/definitions/models.DamageTenantUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated damage ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Lease not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Damage not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Update damage for tenant
      tags:
      - damage
  /tenant/leases/{lease_id}/damages/{damage_id}/fix/:
    put:
      consumes:
      - application/json
      description: Fix a damage for a tenant or owner. When both users have fixed
        the damage, it will be marked as fixed.
      parameters:
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: Damage ID
        in: path
        name: damage_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Fixed damage ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Lease not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Damage not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Fix damage for one user
      tags:
      - damage
  /tenant/leases/{lease_id}/damages/{damage_id}/pictures/:
    post:
      consumes:
      - multipart/form-data
      description: Add pictures to a damage. Only tenant can add pictures.
      parameters:
      - description: Lease ID
        in: path
        name: lease_id
        required: true
        type: string
      - description: Damage ID
        in: path
        name: damage_id
        required: true
        type: string
      - collectionFormat: csv
        description: Files to upload
        in: formData
        items:
          type: file
        name: pictures
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Updated damage ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Lease not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Damage not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Add pictures to damage
      tags:
      - damage
...
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"